cloud_functions/minute-interval-data/requirements.txt
functions-framework==3.*
google-cloud-firestore
flask
google-api-core
google-cloud-storage